Added Forget Password button
Added a button for the user to click if they forgot their password. This leads to a barebones webpage but has routing set and the button is functional. No testing because there is no real functionality. So testing will be added next before the forgot password script is fleshed out.